Change all offsets to be stored in bytes instead of bits
Originally the parser had all offsets specified in bits since we didn't know
that fields in mainframe files are always at least byte-aligned. It is not
possible that any 2 fields share a byte where some of the bits belong to the
first field and some of the bits belong to the second field.