Change semantics of COMP-5
The data items can contain values up to the capacity of the native binary representation
(2, 4 or 8 bytes), rather than being limited to the value implied by the number of nines
in the picture for the item (as is the case for USAGE BINARY data). When numeric data is
moved or stored into a COMP-5 item, truncation occurs at the binary field size rather
than at the COBOL picture size limit. When a COMP-5 item is referenced, the full binary
field size is used in the operation.
(https://www.ibm.com/support/knowledgecenter/en/ssw_ibm_i_72/rzasb/usagcp5.htm)

Use COMP-9 for binary little-endian format